Extract ImageViewController
